+ Version 5.0.0.0 for RunUO 2.7 and ServUO 55+
* Some files have been removed, moved, or renamed.
* It is recommended in this case to completely delete the entire project from your server before installing this version.

+++ Network/ExtendedOPL.cs
@@ -3,7 +3,7 @@
 //   .      __,-; ,'( '/
 //    \.    `-.__`-._`:_,-._       _ , . ``
 //     `:-._,------' ` _,`--` -: `_ , ` ,' :
-//        `---..__,,--'  (C) 2016  ` -'. -'
+//        `---..__,,--'  (C) 2018  ` -'. -'
 //        #  Vita-Nex [http://core.vita-nex.com]  #
 //  {o)xxx|===============-   #   -===============|xxx(o}
 //        #        The MIT License (MIT)          #
@@ -18,16 +18,22 @@ using System.Linq;
 
 using Server;
 using Server.Network;
+
+using VitaNex.Collections;
 #endregion
 
 namespace VitaNex.Network
 {
+	public delegate void OPLQueryValidator(Mobile viewer, IEntity target, ref bool allow);
+
 	/// <summary>
 	///     Provides methods for extending Item and Mobile ObjectPropertyLists by invoking event subscriptions - No more
 	///     GetProperties overrides!
 	/// </summary>
-	public sealed class ExtendedOPL : IList<string>
+	public sealed class ExtendedOPL : IList<string>, IDisposable
 	{
+		private static readonly string[] _EmptyBuffer = new string[0];
+
 		/// <summary>
 		///     Breaks EmptyClilocs every Nth entry. EG: When 5 entries have been added, use the next clilocID
 		/// </summary>
@@ -95,7 +101,12 @@ namespace VitaNex.Network
 		///     Event called when an IEntity based OPL is requested that doesn't match an Item or Mobile
 		/// </summary>
 		public static event Action<IEntity, Mobile, ExtendedOPL> OnEntityOPLRequest;
-		
+
+		/// <summary>
+		///     Event called when an IEntity based OPL is requested
+		/// </summary>
+		public static event OPLQueryValidator OnValidateQuery;
+
 		public static void Init()
 		{
 			if (Initialized)
@@ -119,7 +130,7 @@ namespace VitaNex.Network
 				ReqBatchOplParent.Length,
 				ReqBatchOplParent.Ingame,
 				OnBatchQueryProperties);
-			
+
 			OutParent0xD6 = OutgoingPacketOverrides.GetHandler(0xD6);
 			OutgoingPacketOverrides.Register(0xD6, OnEncode0xD6);
 
@@ -154,20 +165,17 @@ namespace VitaNex.Network
 				mob.GetProperties(opl);
 			}
 
-			var eopl = new ExtendedOPL(opl);
-
-			InvokeOPLRequest(e, v, eopl);
-
-			eopl.Apply();
-
-			//opl = eopl.Opl ?? opl;
+			using (var eopl = new ExtendedOPL(opl))
+			{
+				InvokeOPLRequest(e, v, eopl);
+			}
 
 			opl.Terminate();
 			opl.SetStatic();
-			
+
 			return opl;
 		}
-		
+
 		private static void OnEncode0xD6(NetState state, PacketReader reader, ref byte[] buffer, ref int length)
 		{
 			if (state == null || reader == null || buffer == null || length < 0)
@@ -176,8 +184,11 @@ namespace VitaNex.Network
 			}
 
 			var pos = reader.Seek(0, SeekOrigin.Current);
+
 			reader.Seek(5, SeekOrigin.Begin);
-			Serial serial = reader.ReadInt32();
+
+			var serial = reader.ReadInt32();
+
 			reader.Seek(pos, SeekOrigin.Begin);
 
 			var opl = ResolveOPL(World.FindEntity(serial), state.Mobile);
@@ -195,17 +206,6 @@ namespace VitaNex.Network
 				return;
 			}
 
-			if (OnItemOPLRequest == null && OnMobileOPLRequest == null)
-			{
-				if (ReqBatchOplParent != null)
-				{
-					ReqBatchOplParent.OnReceive(state, pvSrc);
-					return;
-				}
-			}
-
-			var from = state.Mobile;
-
 			var length = pvSrc.Size - 3;
 
 			if (length < 0 || (length % 4) != 0)
@@ -215,28 +215,15 @@ namespace VitaNex.Network
 
 			var count = length / 4;
 
+			Serial s;
+
 			for (var i = 0; i < count; ++i)
 			{
-				Serial s = pvSrc.ReadInt32();
+				s = pvSrc.ReadInt32();
 
-				if (s.IsMobile)
+				if (s.IsValid)
 				{
-					var m = World.FindMobile(s);
-
-					if (m != null && from.CanSee(m) && Utility.InUpdateRange(from, m))
-					{
-						SendPropertiesTo(from, m);
-					}
-				}
-				else if (s.IsItem)
-				{
-					var item = World.FindItem(s);
-
-					if (item != null && !item.Deleted && from.CanSee(item) &&
-						Utility.InUpdateRange(from.Location, item.GetWorldLocation()))
-					{
-						SendPropertiesTo(from, item);
-					}
+					HandleQueryProperties(state.Mobile, World.FindEntity(s));
 				}
 			}
 		}
@@ -248,32 +235,55 @@ namespace VitaNex.Network
 				return;
 			}
 
-			var from = state.Mobile;
-			Serial s = pvSrc.ReadInt32();
+			var serial = (Serial)pvSrc.ReadInt32();
 
-			if (s.IsMobile)
+			if (serial.IsValid)
 			{
-				var m = World.FindMobile(s);
+				HandleQueryProperties(state.Mobile, World.FindEntity(serial));
+			}
+		}
 
-				if (m != null && from.CanSee(m) && Utility.InUpdateRange(from, m))
+		private static void HandleQueryProperties(Mobile viewer, IEntity e)
+		{
+			if (viewer == null || viewer.Deleted || e == null || e.Deleted || !viewer.CanSee(e))
+			{
+				return;
+			}
+
+			if (OnValidateQuery != null)
+			{
+				var allow = true;
+
+				OnValidateQuery(viewer, e, ref allow);
+
+				if (!allow)
 				{
-					SendPropertiesTo(from, m);
+					return;
 				}
 			}
-			else if (s.IsItem)
-			{
-				var item = World.FindItem(s);
 
-				if (item != null && !item.Deleted && from.CanSee(item) &&
-					Utility.InUpdateRange(from.Location, item.GetWorldLocation()))
+			if (e is Mobile)
+			{
+				var m = (Mobile)e;
+
+				if (Utility.InUpdateRange(viewer, m))
 				{
-					SendPropertiesTo(from, item);
+					SendPropertiesTo(viewer, m);
+				}
+			}
+			else if (e is Item)
+			{
+				var item = (Item)e;
+
+				if (Utility.InUpdateRange(viewer, item.GetWorldLocation()))
+				{
+					SendPropertiesTo(viewer, item);
 				}
 			}
 		}
 
 		/// <summary>
-		///     Forces the comilation of a new Mobile based ObjectPropertyList and sends it to the specified Mobile
+		///     Forces the compilation of a new Mobile based ObjectPropertyList and sends it to the specified Mobile
 		/// </summary>
 		/// <param name="to">Mobile viewer, the Mobile viewing the OPL</param>
 		/// <param name="m">Mobile owner, the Mobile which owns the OPL</param>
@@ -293,7 +303,7 @@ namespace VitaNex.Network
 		}
 
 		/// <summary>
-		///     Forces the comilation of a new Item based ObjectPropertyList and sends it to the specified Mobile
+		///     Forces the compilation of a new Item based ObjectPropertyList and sends it to the specified Mobile
 		/// </summary>
 		/// <param name="to">Mobile viewer, the Mobile viewing the OPL</param>
 		/// <param name="item"></param>
@@ -335,38 +345,111 @@ namespace VitaNex.Network
 			}
 		}
 
-		public static void AddTo(ObjectPropertyList opl, string line, params object[] args)
+		public static void AddTo(ObjectPropertyList opl, params object[] args)
 		{
-			if (args != null)
+			if (opl == null || args.IsNullOrEmpty())
 			{
-				line = String.Format(line, args);
+				return;
 			}
 
-			AddTo(opl, new[] {line});
+			using (var o = new ExtendedOPL(opl))
+			{
+				foreach (var a in args)
+				{
+					if (a != null)
+					{
+						o.Add(a.ToString());
+					}
+					else
+					{
+						o.Add(String.Empty);
+					}
+				}
+			}
+		}
+
+		public static void AddTo(ObjectPropertyList opl, IEnumerable<object> args)
+		{
+			if (opl == null || args == null)
+			{
+				return;
+			}
+
+			using (var o = new ExtendedOPL(opl))
+			{
+				foreach (var a in args)
+				{
+					if (a != null)
+					{
+						o.Add(a.ToString());
+					}
+					else
+					{
+						o.Add(String.Empty);
+					}
+				}
+			}
+		}
+
+		public static void AddTo(ObjectPropertyList opl, string line, params object[] args)
+		{
+			if (opl == null)
+			{
+				return;
+			}
+
+			using (var o = new ExtendedOPL(opl))
+			{
+				if (!args.IsNullOrEmpty())
+				{
+					o.Add(line, args);
+				}
+				else
+				{
+					o.Add(line);
+				}
+			}
 		}
 
 		public static void AddTo(ObjectPropertyList opl, string[] lines)
 		{
-			if (opl != null)
+			if (opl == null || lines.IsNullOrEmpty())
 			{
-				new ExtendedOPL(opl, lines).Apply();
+				return;
+			}
+
+			using (var o = new ExtendedOPL(opl))
+			{
+				o.AddRange(lines);
 			}
 		}
-		
+
 		private List<string> _Buffer;
 
-		public int Count { get { return _Buffer.Count; } }
+		bool ICollection<string>.IsReadOnly { get { return _Buffer.IsNullOrEmpty(); } }
 
-		public bool IsReadOnly { get { return false; } }
+		public int Count { get { return _Buffer != null ? _Buffer.Count : 0; } }
 
-		public string this[int index] { get { return _Buffer[index]; } set { _Buffer[index] = value; } }
+		public string this[int index]
+		{
+			get { return _Buffer != null ? _Buffer[index] : null; }
+			set
+			{
+				if (_Buffer != null)
+				{
+					_Buffer[index] = value;
+				}
+			}
+		}
 
 		/// <summary>
 		///     Gets or sets the underlying ObjectPropertyList
 		/// </summary>
 		public ObjectPropertyList Opl { get; set; }
-		
+
 		public int LineBreak { get; set; }
+
+		public bool IsDisposed { get; private set; }
 
 		/// <summary>
 		///     Create with pre-defined OPL
@@ -374,7 +457,7 @@ namespace VitaNex.Network
 		/// <param name="opl">ObjectPropertyList object to wrap and extend</param>
 		public ExtendedOPL(ObjectPropertyList opl)
 		{
-			_Buffer = new List<string>();
+			_Buffer = ListPool<string>.AcquireObject();
 
 			Opl = opl;
 			LineBreak = ClilocBreak;
@@ -387,7 +470,8 @@ namespace VitaNex.Network
 		/// <param name="capacity">Capacity of the extension</param>
 		public ExtendedOPL(ObjectPropertyList opl, int capacity)
 		{
-			_Buffer = new List<string>(capacity);
+			_Buffer = ListPool<string>.AcquireObject();
+			_Buffer.Capacity = capacity;
 
 			Opl = opl;
 			LineBreak = ClilocBreak;
@@ -400,7 +484,8 @@ namespace VitaNex.Network
 		/// <param name="list">Pre-defined list to append to the specified OPL</param>
 		public ExtendedOPL(ObjectPropertyList opl, IEnumerable<string> list)
 		{
-			_Buffer = new List<string>(list);
+			_Buffer = ListPool<string>.AcquireObject();
+			_Buffer.AddRange(list);
 
 			Opl = opl;
 			LineBreak = ClilocBreak;
@@ -408,17 +493,53 @@ namespace VitaNex.Network
 
 		~ExtendedOPL()
 		{
-			Opl = null;
+			Dispose(true);
+		}
 
-			_Buffer.Free(true);
-			_Buffer = null;
+		void IDisposable.Dispose()
+		{
+			Dispose(false);
+		}
+
+		private void Dispose(bool gc)
+		{
+			if (IsDisposed)
+			{
+				return;
+			}
+
+			try
+			{
+				if (!gc)
+				{
+					GC.SuppressFinalize(this);
+				}
+			}
+			catch
+			{ }
+
+			try
+			{
+				if (Opl != null)
+				{
+					Flush();
+
+					Opl = null;
+				}
+			}
+			catch
+			{ }
+
+			ObjectPool.Free(ref _Buffer);
+
+			IsDisposed = true;
 		}
 
 		public bool NextEmpty(out ClilocInfo info)
 		{
 			info = null;
 
-			if (Opl == null)
+			if (IsDisposed || Opl == null)
 			{
 				return false;
 			}
@@ -430,7 +551,11 @@ namespace VitaNex.Network
 				if (!Opl.Contains(index))
 				{
 					info = ClilocLNG.NULL.Lookup(index);
-					return true;
+
+					if (info != null)
+					{
+						return true;
+					}
 				}
 			}
 
@@ -439,9 +564,17 @@ namespace VitaNex.Network
 
 		public void Flush()
 		{
+			if (IsDisposed)
+			{
+				return;
+			}
+
 			Apply();
 
-			_Buffer.Free(true);
+			if (_Buffer != null)
+			{
+				_Buffer.Free(true);
+			}
 		}
 
 		/// <summary>
@@ -449,9 +582,18 @@ namespace VitaNex.Network
 		/// </summary>
 		public void Apply()
 		{
-			if (Opl == null || _Buffer.Count == 0)
+			if (IsDisposed)
 			{
-				_Buffer.Clear();
+				return;
+			}
+
+			if (Opl == null || _Buffer == null || _Buffer.Count == 0)
+			{
+				if (_Buffer != null)
+				{
+					_Buffer.Clear();
+				}
+
 				return;
 			}
 
@@ -478,14 +620,14 @@ namespace VitaNex.Network
 				{
 					var s = _Buffer[i];
 
-					if (i > 0)
-					{
-						final += '\n';
-					}
-
 					if (String.IsNullOrWhiteSpace(s))
 					{
 						s = " ";
+					}
+
+					if (i > 0)
+					{
+						final += '\n';
 					}
 
 					final += s;
@@ -503,19 +645,30 @@ namespace VitaNex.Network
 
 				_Buffer.RemoveRange(0, take);
 
-				Opl.Add(info.Index, info.ToString(final));
+				if (!String.IsNullOrEmpty(final))
+				{
+					Opl.Add(info.Index, info.ToString(final));
+				}
 			}
 
-			_Buffer.Clear();
+			Clear();
 		}
 
 		public void Add(string format, params object[] args)
 		{
-			_Buffer.Add(String.Format(format ?? String.Empty, args));
+			if (_Buffer != null)
+			{
+				_Buffer.Add(String.Format(format ?? String.Empty, args));
+			}
 		}
 
 		public void Add(int number, params object[] args)
 		{
+			if (_Buffer == null)
+			{
+				return;
+			}
+
 			var info = ClilocLNG.NULL.Lookup(number);
 
 			if (info != null)
@@ -524,53 +677,94 @@ namespace VitaNex.Network
 			}
 		}
 
+		public void Add(ObjectPropertyList opl)
+		{
+			Add(opl, ClilocLNG.NULL);
+		}
+
+		public void Add(ObjectPropertyList opl, ClilocLNG lng)
+		{
+			if (opl != null && opl != Opl && opl.Entity != Opl.Entity)
+			{
+				AddRange(opl.GetAllLines(lng));
+			}
+		}
+
 		public void AddRange(IEnumerable<string> lines)
 		{
-			_Buffer.AddRange(lines.Select(line => line ?? String.Empty));
+			if (_Buffer != null)
+			{
+				_Buffer.AddRange(lines.Select(line => line ?? String.Empty));
+			}
 		}
 
 		public void Add(string line)
 		{
-			_Buffer.Add(line ?? String.Empty);
+			if (_Buffer != null)
+			{
+				_Buffer.Add(line ?? String.Empty);
+			}
 		}
 
 		public bool Remove(string line)
 		{
-			return _Buffer.Remove(line ?? String.Empty);
+			if (_Buffer != null)
+			{
+				return _Buffer.Remove(line ?? String.Empty);
+			}
+
+			return false;
 		}
 
 		public int RemoveAll(Predicate<string> match)
 		{
-			return _Buffer.RemoveAll(match);
+			if (_Buffer != null && match != null)
+			{
+				return _Buffer.RemoveAll(match);
+			}
+
+			return 0;
 		}
 
 		public void RemoveRange(int index, int count)
 		{
-			_Buffer.RemoveRange(index, count);
+			if (_Buffer != null)
+			{
+				_Buffer.RemoveRange(index, count);
+			}
 		}
 
 		public void RemoveAt(int index)
 		{
-			_Buffer.RemoveAt(index);
-		}
-
-		public int IndexOf(string line)
-		{
-			return _Buffer.IndexOf(line ?? String.Empty);
+			if (_Buffer != null)
+			{
+				_Buffer.RemoveAt(index);
+			}
 		}
 
 		public void Insert(int index, string line)
 		{
-			_Buffer.Insert(index, line ?? String.Empty);
+			if (_Buffer != null)
+			{
+				_Buffer.Insert(index, line ?? String.Empty);
+			}
 		}
 
 		public void Insert(int index, string format, params object[] args)
 		{
-			_Buffer.Insert(index, String.Format(format ?? String.Empty, args));
+			if (_Buffer != null)
+			{
+				_Buffer.Insert(index, String.Format(format ?? String.Empty, args));
+			}
 		}
 
 		public void Insert(int index, int number, params object[] args)
 		{
+			if (_Buffer == null)
+			{
+				return;
+			}
+
 			var info = ClilocLNG.NULL.Lookup(number);
 
 			if (info != null)
@@ -579,29 +773,50 @@ namespace VitaNex.Network
 			}
 		}
 
+		public int IndexOf(string line)
+		{
+			if (_Buffer != null)
+			{
+				return _Buffer.IndexOf(line ?? String.Empty);
+			}
+
+			return -1;
+		}
+
 		public bool Contains(string line)
 		{
-			return _Buffer.Contains(line ?? String.Empty);
+			return _Buffer != null && _Buffer.Contains(line ?? String.Empty);
 		}
 
 		public void Clear()
 		{
-			_Buffer.Clear();
+			if (_Buffer != null)
+			{
+				_Buffer.Clear();
+			}
 		}
 
 		public void CopyTo(string[] array, int arrayIndex)
 		{
-			_Buffer.CopyTo(array, arrayIndex);
+			if (_Buffer != null)
+			{
+				_Buffer.CopyTo(array, arrayIndex);
+			}
 		}
 
 		public IEnumerator<string> GetEnumerator()
 		{
-			return _Buffer.GetEnumerator();
+			if (_Buffer != null)
+			{
+				return _Buffer.GetEnumerator();
+			}
+
+			return _EmptyBuffer.GetEnumerator<string>();
 		}
 
 		IEnumerator IEnumerable.GetEnumerator()
 		{
-			return _Buffer.GetEnumerator();
+			return GetEnumerator();
 		}
 	}
 }
